+++ ui/src/App.tsx
@@ -1,72 +1,24 @@
 import { MantineProvider } from "@mantine/core";
 import { Root } from "@/components/Root";
-import { OutputData } from "@editorjs/editorjs";
-import { useEffect, useRef, useState } from "react";
+import { lazy, Suspense, useRef } from "react";
 import {
   BrowserRouter as Router,
   Route,
   Routes,
   Navigate,
-  useParams,
 } from "react-router-dom";
-import { Editor } from "./components/Editor";
 import {
   EditorContext,
   EditorInstanceContainer,
 } from "./components/Editor/Context";
-import { SpinnerIcon } from "./components/SpinnerIcon";
 import { Nav } from "./layout/Nav";
-import { DefaultClient } from "./services/client";
-import { PlaygroundPage } from "./Playground";
 import { QueryClient, QueryClientProvider } from "react-query";
-import { CollectionsPage } from "./components/Collections";
-
-function ContentPage() {
-  const { id } = useParams();
-
-  const [isLoading, setLoading] = useState(true);
-  const [initialValue, setInitialValue] = useState<OutputData | undefined>();
-
-  useEffect(() => {
-    async function work() {
-      setLoading(true);
-      try {
-        const resp = await DefaultClient.API.notebookGetSharedData({
-          gistId: id!,
-        });
-        const data = JSON.parse(resp.data.data);
-        setInitialValue(data);
-      } finally {
-        setLoading(false);
-      }
-    }
-    if (id) {
-      work();
-    } else {
-      setLoading(false);
-    }
-  }, [id]);
-
-  return (
-    <div className="bg-slate-100">
-      <div className="w-[1150px] mx-auto">
-        <div className="pt-16 w-[850px] bg-white shadow-xl min-h-screen">
-          {isLoading && (
-            <div className="w-[650px] mx-auto">
-              <div className="flex items-center">
-                <SpinnerIcon className="mr-2 w-4 h-4 text-gray-200 animate-spin fill-rose-500" />
-                <span className="text-gray-600">Loading notebook...</span>
-              </div>
-            </div>
-          )}
-          {!isLoading && <Editor defaultValue={initialValue} />}
-        </div>
-      </div>
-    </div>
-  );
-}
 
 const queryClient = new QueryClient();
+
+const ContentPage = lazy(() => import("./ContentPage"));
+const PlaygroundPage = lazy(() => import("./Playground"));
+const CollectionsPage = lazy(() => import("./components/Collections"));
 
 export default function App() {
   // A globally available editor.js instance
@@ -87,9 +39,30 @@ export default function App() {
               <div className="h-screen">
                 <Nav />
                 <Routes>
-                  <Route path="/create" element={<ContentPage />} />
-                  <Route path="/gist/:id" element={<ContentPage />} />
-                  <Route path="/playground" element={<PlaygroundPage />} />
+                  <Route
+                    path="/create"
+                    element={
+                      <Suspense fallback={<></>}>
+                        <ContentPage />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="/gist/:id"
+                    element={
+                      <Suspense fallback={<></>}>
+                        <ContentPage />
+                      </Suspense>
+                    }
+                  />
+                  <Route
+                    path="/playground"
+                    element={
+                      <Suspense fallback={<></>}>
+                        <PlaygroundPage />
+                      </Suspense>
+                    }
+                  />
                   <Route path="/" element={<CollectionsPage />}></Route>
                   <Route
                     path="*"
